Embed swagger API before packaging master template
